Booking Implementation

+++ src/main/java/com/lifeassist/repository/CaregiverRepository.java
@@ -4,9 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.lifeassist.entity.CaregiverDetails;
-import com.lifeassist.entity.CaretakerDetails;
 @Repository
-public interface CaregiverRepository extends JpaRepository<CaregiverDetails, Integer>{
+public interface CaregiverRepository extends JpaRepository<CaregiverDetails, Long>{
 
 
 }
