+++ src/main/java/com/lifeassist/repository/GuardianRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lifeassist.entity.GuardianDetails;
 
-public interface GuardianRepository extends JpaRepository<GuardianDetails, Integer>{
+public interface GuardianRepository extends JpaRepository<GuardianDetails, Long>{
 
 }
